+++ src/components/atoms/form/FormContainer/style.module.css
@@ -1,0 +1,10 @@
+.container {
+  margin-top: 80px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  border: 1px solid #707070;
+  border-radius: 20px;
+  padding: 20px 50px;
+}
