+++ src/components/organisms/signUp/SignUpForm/style.module.css
@@ -5,14 +5,7 @@
 }
 
 .container {
-  margin-top: 80px;
-  display: flex;
-  justify-content: center;
-  align-items: left;
-  flex-direction: column;
-  border: 1px solid #707070;
-  border-radius: 20px;
-  padding: 20px 50px;
+  align-items: unset;
 }
 
 .label {
