+++ src/components/pages/SignIn/style.module.css
@@ -9,17 +9,6 @@
   align-items: center;
 }
 
-.form-container {
-  margin-top: 80px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  border: 1px solid #707070;
-  border-radius: 20px;
-  padding: 20px 50px;
-}
-
 .form-field {
   margin: 15px 0;
 }
